refactor(journal): move SegmenteDescriptorReader to own class
